refactor: exception 패키지 위치 변경
domain과 같은 레벨로 변경
