+++ src/main/java/roomescape/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package roomescape.global.exception;
+package roomescape.exception;
 
 import org.springframework.http.HttpStatus;
