+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -5,8 +5,8 @@ import java.util.ArrayList;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.global.exception.ErrorCode;
-import roomescape.global.exception.RoomescapeException;
+import roomescape.exception.ErrorCode;
+import roomescape.exception.RoomescapeException;
 import roomescape.reservation.Reservation;
 import roomescape.reservation.TotalReservation;
 import roomescape.reservation.dao.ReservationDao;
