+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -19,8 +19,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.global.exception.ErrorCode;
-import roomescape.global.exception.RoomescapeException;
+import roomescape.exception.ErrorCode;
+import roomescape.exception.RoomescapeException;
 import roomescape.reservation.Reservation;
 import roomescape.reservation.TotalReservation;
 import roomescape.reservation.dao.ReservationDao;
